+++ Assets/QA/Tests/XRDeviceActions/_SharedScripts/Vector3ControlActionStatus.cs
@@ -18,7 +18,7 @@ public class Vector3ControlActionStatus : MonoBehaviour
         vector3Action.Enable();
         vector3Action.performed += UpdateVector3;
         vector3Action.started += UpdateVector3;
-        vector3Action.cancelled += UpdateVector3;
+        vector3Action.canceled += UpdateVector3;
 
         ReadOnlyArray<InputControl> controls = vector3Action.controls;
         for (int i = 0; i < controls.Count; i++)
@@ -41,7 +41,7 @@ public class Vector3ControlActionStatus : MonoBehaviour
         vector3Action.Disable();
         vector3Action.performed -= UpdateVector3;
         vector3Action.started -= UpdateVector3;
-        vector3Action.cancelled -= UpdateVector3;
+        vector3Action.canceled -= UpdateVector3;
     }
 
     private void UpdateVector3(InputAction.CallbackContext context)