+++ Packages/com.unity.inputsystem/InputSystem/Plugins/UI/TouchModel.cs
@@ -84,7 +84,7 @@ namespace UnityEngine.InputSystem.Plugins.UI
                     if (value == PointerPhase.Began)
                         selectDelta |= ButtonDeltaState.Pressed;
 
-                    if (value == PointerPhase.Ended || value == PointerPhase.Cancelled)
+                    if (value == PointerPhase.Ended || value == PointerPhase.Canceled)
                         selectDelta |= ButtonDeltaState.Released;
 
                     m_SelectPhase = value;
@@ -126,7 +126,7 @@ namespace UnityEngine.InputSystem.Plugins.UI
 
             m_Position = deltaPosition = Vector2.zero;
 
-            m_SelectPhase = PointerPhase.Cancelled;
+            m_SelectPhase = PointerPhase.Canceled;
             changedThisFrame = false;
             selectDelta = ButtonDeltaState.NoChange;